requirements.txt
aiodns==1.1.1
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
autopep8==1.4.4
ccxt==1.18.555
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
idna==2.8
multidict==4.5.2
pycares==3.0.0
pycodestyle==2.5.0
pycparser==2.19
python-dotenv==0.10.2
requests==2.22.0
urllib3==1.25.2
yarl==1.1.0
